Update api.rs
Updated Api struct to have a reqwest::Client member, which will be more efficient in the long run.

Added a function, new, to more easily instantiate the struct; Returns an Error if the auth object is not valid.

Refactored buildUrl to only build the url, removing the parameters parameter entirely.

Added asynchronous generic function get, since the process of executing the HTTP request will be the same for every endpoint.

Refactored function getPlayerSummaries to use the new get<T> function.

Also built a post<T> function but the Steam Web API disallows POST requests, so it's not useful at the moment.