Update app.rs
Updated the api ref to attempt to generate a new instance of SteamApi with the steamAuth property, if it is some value.

Updated the "Update" button click handler to also assign the new auth instance into the api ref.

The above updates made it easy to simplify the "Run" click handler. Also, prints on future.value None, just to be explicit.