Update api.rs
Added doc comments for cacheGameIcons and cacheProfileAvatar.

Renamed parameter refresh to force in cacheProfileAvatar.
Added parameter force to cacheGameIcons.
Updated cacheGameIcons to return the list of games for which it failed to retrieve icons, if any; Also, removed debug/development print statements.

Updated app.rs to reflect the changes in api.rs:
- Added print statements to the click handler for Get Owned Games to see the results of cacheGameIcons.