Update gamelist.rs
Updated GameList component to output a containing div with id.

Added inline property, refresh, to both GameList and Game which controls a custom "refresh" attribute, allowing both components to be forced to redraw at will.

Updated Game component to only output the img tag when the game's icon exists in the local cache.