Update main.rs
Updated to be a bit more specific with launching the application.

Added a crude workaround to be able to load the CSS styles properly when using dx serve; Will need to find a better solution to this at some point.

Added configuration guard to prevent Windows from opening a Command Prompt when starting the application in a Release build.