Update game.rs
Updated the means of showing/hiding the achievement list to be based in CSS, so that animated transitions are more easily achieved.

Moved the click handlers from the steam api buttons (retrieving game schema and player achievements) into the click handler for the header, so both operations are triggered when the user clicks on the header and that game has no achievements.