Update main.rs
Using tokio::sync::mpsc communication channels for frontend->dispatcher and dispatcher->frontend communication.

Added transmitter() and receiver() to provide access on the frontend to static instances of UnboundedSender and UnboundedReceiver respectively; The values for the static Mutexes are set in main().

Added userData() to provide static access to a single Mutex<User> reference.

Added transmit() and transmitMultiple() as convenience methods for utilizing the transmitter() provided static reference to transmit one or more ApiCommand instances to the Dispatcher; One line vs 7-10 lines.

Updated main() to spawn a tokio task which creates a new Dispatcher instance and starts it listening.