Create app.rs
Added App node to be the root node of the whole application; Handles generating the GameNode instances based on the User data from the AppData node.

Updated GameNode to allow setting its current Game instance and refactored the existing functionality to be driven by that Game value.

Updated RetroAchievementNode to expect its id value to be a string instead of a number; Also, Points is no longer optional in a RetroAchievement instance.