Update achievement.rs, constants.rs, achievement.rs, and 1 more file
Updated the data::retroachievements::game::Game struct to contain the three values for distinct players; the casual and hardcore values might be unnecessary but they're there for now.
Added unlockedPercent function to data::retroachievements::achievement::Achievement to simplify getting the current percentage of players who have unlocked the achievement for the given AchievementMode.
Updated the AchievementElement to not only display the counts and percentages of Casual and Hardcore players who have unlocked the achievement but also more closely imitate the website's representation of this information.