Update content.rs, game.rs, list.rs, and 3 more files
Redesigned the RetroAchievements game list; rather than using the Accordion and nesting VirtualScrollView components, just have the game list effectively be a button and replace the list entirely when a game is chosen. Also, updated the game list with RetroAchievements' more detailed information and to be a bit more stylish.
Added a Back button to the RetroAchievements GameElement, to return to the game list.
Update the RetroAchievements Game data struct to sort by name first, then by game system.